Prevent unwanted brutal file deletiong by sanity checking begore the 'rm -rf' command
If the 'INITRAMFS_MODULES_DIR' would contain only '/' for some reason, we would have
a problem, therefore checking that the path is set correctly in the variable.


Contributes to: EMP-607.

+++ build_for_ultimaker.sh
@@ -169,7 +169,7 @@ initramfs_prepare()
         busybox_get "${INITRAMFS_DST_DIR}"
     fi
 
-    if [ -d "${INITRAMFS_MODULES_DIR}" ]; then
+    if [ -d "${INITRAMFS_MODULES_DIR}" ] && [ -z "${INITRAMFS_MODULES_DIR##*/initramfs/lib/modules*}" ]; then
         rm -rf "${INITRAMFS_MODULES_DIR}"
     fi
